# tests/snapshots/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ConfirmDecomposeBootstrapWorkflowResponse": {
        "properties": {
          "approved": {
            "title": "Approved",
            "type": "boolean"
          },
          "confirmation_status": {
            "title": "Confirmation Status",
            "type": "string"
          },
          "confirmation_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Token"
          },
          "confirmed_by": {
            "title": "Confirmed By",
            "type": "string"
          },
          "modules": {
            "items": {
              "type": "string"
            },
            "title": "Modules",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "workitems": {
            "items": {
              "$ref": "#/components/schemas/WorkItem"
            },
            "title": "Workitems",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "approved",
          "confirmation_status",
          "confirmed_by"
        ],
        "title": "ConfirmDecomposeBootstrapWorkflowResponse",
        "type": "object"
      },
...
      "DecomposeBootstrapAdvanceLoopResponse": {
        "properties": {
          "action_status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Action Status Counts",
            "type": "object"
          },
          "action_taken_sequence": {
            "items": {
              "type": "string"
            },
            "title": "Action Taken Sequence",
            "type": "array"
          },
          "final_status": {
            "$ref": "#/components/schemas/DecomposeBootstrapAggregateStatusResponse"
          },
          "halted_reason": {
            "title": "Halted Reason",
            "type": "string"
          },
          "last_action_taken": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Action Taken"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/DecomposeBootstrapAdvanceResponse"
            },
            "title": "Steps",
            "type": "array"
          },
          "steps_executed": {
            "default": 0,
            "title": "Steps Executed",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "halted_reason",
          "final_status"
        ],
        "title": "DecomposeBootstrapAdvanceLoopResponse",
        "type": "object"
      },
...
      "DecomposeBootstrapAdvanceResponse": {
        "properties": {
          "action_status": {
            "title": "Action Status",
            "type": "string"
          },
          "action_taken": {
            "title": "Action Taken",
            "type": "string"
          },
          "confirmation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfirmDecomposeBootstrapWorkflowResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "execute": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecuteWorkflowRunResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecomposeBootstrapPreviewResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status_after": {
            "$ref": "#/components/schemas/DecomposeBootstrapAggregateStatusResponse"
          },
          "status_before": {
            "$ref": "#/components/schemas/DecomposeBootstrapAggregateStatusResponse"
          }
        },
        "required": [
          "run_id",
          "action_taken",
          "action_status",
          "status_before",
          "status_after"
        ],
        "title": "DecomposeBootstrapAdvanceResponse",
        "type": "object"
      },
      "DecomposeBootstrapAggregateStatusResponse": {
        "properties": {
          "bootstrap_finished": {
            "default": false,
            "title": "Bootstrap Finished",
            "type": "boolean"
          },
          "bootstrap_started": {
            "default": false,
            "title": "Bootstrap Started",
            "type": "boolean"
          },
          "confirmation_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Status"
          },
          "decomposition_source": {
            "default": "none",
            "title": "Decomposition Source",
            "type": "string"
          },
          "global_workitem_count": {
            "default": 0,
            "title": "Global Workitem Count",
            "type": "integer"
          },
          "has_decomposition": {
            "default": false,
            "title": "Has Decomposition",
            "type": "boolean"
          },
          "has_pending_confirmation": {
            "default": false,
            "title": "Has Pending Confirmation",
            "type": "boolean"
          },
          "module_workitem_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Module Workitem Counts",
            "type": "object"
          },
          "modules": {
            "items": {
              "type": "string"
            },
            "title": "Modules",
            "type": "array"
          },
          "next_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Action"
          },
          "preview_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Fingerprint"
          },
          "preview_generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Generated At"
          },
          "preview_ready": {
            "default": false,
            "title": "Preview Ready",
            "type": "boolean"
          },
          "preview_stale": {
            "default": false,
            "title": "Preview Stale",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "run_status": {
            "$ref": "#/components/schemas/WorkflowRunStatus"
          },
          "workitem_status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Workitem Status Counts",
            "type": "object"
          },
          "workitem_total": {
            "default": 0,
            "title": "Workitem Total",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "run_status"
        ],
        "title": "DecomposeBootstrapAggregateStatusResponse",
        "type": "object"
      },
...
      "DecomposeBootstrapPreviewResponse": {
        "properties": {
          "cache_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Fingerprint"
          },
          "cache_hit": {
            "default": false,
            "title": "Cache Hit",
            "type": "boolean"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "modules": {
            "items": {
              "type": "string"
            },
            "title": "Modules",
            "type": "array"
          },
          "parallel_groups": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Parallel Groups",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "task_count": {
            "default": 0,
            "title": "Task Count",
            "type": "integer"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/DecomposeBootstrapPreviewTask"
            },
            "title": "Tasks",
            "type": "array"
          },
          "terminal_task_keys": {
            "items": {
              "type": "string"
            },
            "title": "Terminal Task Keys",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "source"
        ],
        "title": "DecomposeBootstrapPreviewResponse",
        "type": "object"
      },
...
      "DiscussionSession": {
        "properties": {
          "budget": {
            "default": 2,
            "minimum": 0.0,
            "title": "Budget",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fingerprint"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "impact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact"
          },
          "opened_by_role": {
            "minLength": 1,
            "title": "Opened By Role",
            "type": "string"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "title": "Options",
            "type": "array"
          },
          "question": {
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "recommendation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation"
          },
          "resolved_by_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By Role"
          },
          "round": {
            "default": 1,
            "minimum": 1.0,
            "title": "Round",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/DiscussionStatus",
            "default": "open"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow_run_id": {
            "minLength": 1,
            "title": "Workflow Run Id",
            "type": "string"
          },
          "workitem_id": {
            "minLength": 1,
            "title": "Workitem Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_run_id",
          "workitem_id",
          "question",
          "opened_by_role"
        ],
        "title": "DiscussionSession",
        "type": "object"
      },
      "DiscussionStatus": {
        "enum": [
          "open",
          "resolved",
          "exhausted",
          "timeout"
        ],
        "title": "DiscussionStatus",
        "type": "string"
      },
      "ExecuteWorkflowRunRequest": {
        "properties": {
          "auto_advance_decompose": {
            "default": true,
            "title": "Auto Advance Decompose",
            "type": "boolean"
          },
          "auto_advance_execute_max_loops": {
            "anyOf": [
              {
                "maximum": 1000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Advance Execute Max Loops"
          },
          "auto_advance_force_refresh_preview": {
            "default": false,
            "title": "Auto Advance Force Refresh Preview",
            "type": "boolean"
          },
          "auto_advance_max_steps": {
            "default": 8,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Auto Advance Max Steps",
            "type": "integer"
          },
          "decompose_confirmation_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decompose Confirmation Token"
          },
          "decompose_confirmed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decompose Confirmed By"
          },
          "decompose_expected_modules": {
            "items": {
              "type": "string"
            },
            "title": "Decompose Expected Modules",
            "type": "array"
          },
          "max_loops": {
            "default": 20,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Max Loops",
            "type": "integer"
          }
        },
        "title": "ExecuteWorkflowRunRequest",
        "type": "object"
      },
      "ExecuteWorkflowRunResponse": {
        "properties": {
          "decompose_auto_advance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecomposeBootstrapAdvanceLoopResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "executed_count": {
            "title": "Executed Count",
            "type": "integer"
          },
          "executed_workitem_ids": {
            "items": {
              "type": "string"
            },
            "title": "Executed Workitem Ids",
            "type": "array"
          },
          "failed_count": {
            "title": "Failed Count",
            "type": "integer"
          },
          "failed_workitem_ids": {
            "items": {
              "type": "string"
            },
            "title": "Failed Workitem Ids",
            "type": "array"
          },
          "remaining_pending_count": {
            "title": "Remaining Pending Count",
            "type": "integer"
          },
          "remaining_ready_count": {
            "title": "Remaining Ready Count",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "run_status": {
            "$ref": "#/components/schemas/WorkflowRunStatus"
          },
          "waiting_approval_count": {
            "title": "Waiting Approval Count",
            "type": "integer"
          },
          "waiting_approval_workitem_ids": {
            "items": {
              "type": "string"
            },
            "title": "Waiting Approval Workitem Ids",
            "type": "array"
          },
          "waiting_discussion_count": {
            "title": "Waiting Discussion Count",
            "type": "integer"
          },
          "waiting_discussion_workitem_ids": {
            "items": {
              "type": "string"
            },
            "title": "Waiting Discussion Workitem Ids",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "run_status",
          "executed_count",
          "failed_count",
          "remaining_ready_count",
          "remaining_pending_count",
          "waiting_discussion_count",
          "waiting_approval_count"
        ],
        "title": "ExecuteWorkflowRunResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InterruptWorkflowRunRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "requested_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By"
          },
          "skip_non_terminal_workitems": {
            "default": true,
            "title": "Skip Non Terminal Workitems",
            "type": "boolean"
          }
        },
        "title": "InterruptWorkflowRunRequest",
        "type": "object"
      },
      "InterruptWorkflowRunResponse": {
        "properties": {
          "interrupt_applied": {
            "default": false,
            "title": "Interrupt Applied",
            "type": "boolean"
          },
          "previous_status": {
            "$ref": "#/components/schemas/WorkflowRunStatus"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "run_status": {
            "$ref": "#/components/schemas/WorkflowRunStatus"
          },
          "skipped_workitem_ids": {
            "items": {
              "type": "string"
            },
            "title": "Skipped Workitem Ids",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "previous_status",
          "run_status"
        ],
        "title": "InterruptWorkflowRunResponse",
        "type": "object"
      },
...
      "ResolveDiscussionRequest": {
        "properties": {
          "decision": {
            "minLength": 1,
            "title": "Decision",
            "type": "string"
          },
          "discussion_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discussion Id"
          },
          "resolved_by": {
            "minLength": 1,
            "title": "Resolved By",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "resolved_by"
        ],
        "title": "ResolveDiscussionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkItem": {
        "properties": {
          "assignee_agent": {
            "default": "auto-agent",
            "minLength": 1,
            "title": "Assignee Agent",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "title": "Depends On",
            "type": "array"
          },
          "discussion_budget": {
            "default": 2,
            "minimum": 0.0,
            "title": "Discussion Budget",
            "type": "integer"
          },
          "discussion_timeout_seconds": {
            "default": 120,
            "minimum": 1.0,
            "title": "Discussion Timeout Seconds",
            "type": "integer"
          },
          "discussion_used": {
            "default": 0,
            "minimum": 0.0,
            "title": "Discussion Used",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "loop_guard_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loop Guard Fingerprint"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "module_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module Key"
          },
          "priority": {
            "default": 3,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Priority",
            "type": "integer"
          },
          "requires_approval": {
            "default": false,
            "title": "Requires Approval",
            "type": "boolean"
          },
          "role": {
            "minLength": 1,
            "title": "Role",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/WorkItemStatus",
            "default": "pending"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow_run_id": {
            "minLength": 1,
            "title": "Workflow Run Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_run_id",
          "role"
        ],
        "title": "WorkItem",
        "type": "object"
      },
      "WorkItemStatus": {
        "enum": [
          "pending",
          "ready",
          "running",
          "needs_discussion",
          "waiting_approval",
          "failed",
          "succeeded",
          "skipped"
        ],
        "title": "WorkItemStatus",
        "type": "string"
      },
...
      "WorkflowRunStatus": {
        "enum": [
          "planning",
          "running",
          "waiting_approval",
          "blocked",
          "failed",
          "succeeded",
          "canceled"
        ],
        "title": "WorkflowRunStatus",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "WhereCode Control Center",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v3/workflows/runs/{run_id}/execute": {
      "post": {
        "operationId": "execute_workflow_run_v3_workflows_runs__run_id__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteWorkflowRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteWorkflowRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Workflow Run"
      }
    },
...
    "/v3/workflows/runs/{run_id}/interrupt": {
      "post": {
        "operationId": "interrupt_workflow_run_v3_workflows_runs__run_id__interrupt_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterruptWorkflowRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterruptWorkflowRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Interrupt Workflow Run"
      }
    },
...
    "/v3/workflows/workitems/{workitem_id}/discussion/resolve": {
      "post": {
        "operationId": "resolve_workitem_discussion_v3_workflows_workitems__workitem_id__discussion_resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "workitem_id",
            "required": true,
            "schema": {
              "title": "Workitem Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDiscussionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscussionSession"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Workitem Discussion"
      }
    },
    "/v3/workflows/workitems/{workitem_id}/discussions": {
      "get": {
        "operationId": "list_workitem_discussions_v3_workflows_workitems__workitem_id__discussions_get",
        "parameters": [
          {
            "in": "path",
            "name": "workitem_id",
            "required": true,
            "schema": {
              "title": "Workitem Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DiscussionSession"
                  },
                  "title": "Response List Workitem Discussions V3 Workflows Workitems  Workitem Id  Discussions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Workitem Discussions"
      }
    },
...
  }
}